backend/pyproject.toml
[tool.poetry]
name = "licitacao-hub-api"
version = "0.1.0"
description = "API Backend for Licitação Hub"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.scripts]
start = "app.main:start"
cli = "app.core.cli:app"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
alembic = "^1.13.1"
psycopg = {extras = ["binary", "pool"], version = "^3.2.7"}
pydantic = {extras = ["email"], version = "^2.6.1"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.7"
httpx = {extras = ["http2"], version = "^0.28.1"}
redis = {extras = ["hiredis"], version = "^5.0.1"}
structlog = "^24.1.0"
python-dotenv = "^1.0.1"
typer = {extras = ["all"], version = "^0.9.0"}
jinja2 = "^3.1.2"
pypdf = "^4.0.1"
pyyaml = "^6.0.1"
typing-extensions = "^4.9.0"
dnspython = "^2.6.0"
certifi = "^2024.2.2"
cryptography = "^42.0.2"
greenlet = "^3.0.3"
h11 = "^0.14.0"
watchfiles = "^0.21.0"
rich = "^13.7.0"
rich-toolkit = "^0.1.0"
shellingham = "^1.5.4"
sniffio = "^1.3.1"
pypdf2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5,<0.24.0"
pytest-httpx = "0.35.0"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.2.2"
mypy = "^1.8.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "UP", "N", "ANN", "S", "B", "A", "C4", "T20", "PT", "RET"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.isort]
case-sensitive = true
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
